Append function + array formatting
added in the append function, php won’t overwrite existing arrays but
add on to the file.

Still have not solved the formatting for the form itself, can’t seem to
figure out how to add a colon in or remove inverted commas where i
don’t want it to.